tests/submission/Google/video/controls/controls_reflects_controls.html

<!doctype html>
<html>
 <head>
  <title>video.controls - reflection test</title>
  <script src="../../../../resources/testharness.js"></script>
  <script src="../../../../resources/testharnessreport.js"></script>
 </head>
 <body>
  <p><a href="http://dev.w3.org/html5/spec/Overview.html#dom-media-controls">spec reference</a></p>
  <video id="video" controls="controls">
  </video>
  <div id="log"></div>
  <script>
test(function() {
 assert_true(
  document.getElementById("video").controls,
  "videoElement.controls reflects true when present");
});
  </script>
 </body>
</html>
